Update Instructions to Install emacs-mac with homebrew
Updated installation to include the --with-native-compilation flag for compatibility with emacs@29.1, which is installed now by default. (The current instructions will create a broken config by default)
Updated the command for adding an Emacs.app link to /Applications to a more Mac-ish solution. Added comment explaining that command is optional and linked to other options.