bump: :emacs vc
magit/magit@97a95f70079b -> magit/magit@48818355728c

The pinned version of `magit` in `:tools magit` needs be the same
as that of `git-commit` in `:emacs vc`.

`straight.el` prepares the repo only once, and if `git-commit` is
processed first, its pinned version wins out and the more recent
pinning of `magit` is ignored.

As noted in #7363, processing order may be non-deterministic, so the
inconsistent pinning may not be apparent on every system.

Fix: #7363
Amend: #7277
Amend: 7c63b353d290