fix(lsp): eglot: prefer built-in eglot
Eglot pulled installed via straight caused issues with calls to
`consult-imenu` for my emacs installation of version 29.1.
The issue caused symbols to not be categorized according to class,
function, method, etc. for all languages that were using eglot as lsp.
Using the built-in version of eglot for solves this issue.

Signed-off-by: Tobias Zillig <tobizillig@gmail.com>