fix(vertico) set root user for consult-dir containers
The tramp string used to connect to the container uses the first
'container-runtime ps' argument as a username. However, the first
argument that is returned is the container id, not a username.

Instead of inspecting the container to find the user, just default to
'root' which should work for most cases.