fix: scratch buffer initializing too early
This is a daemon-only issue.
The scratch buffer is initialized very early
in the startup process. However, some minor
modes (such as `flycheck-posframe-mode`) and
other things that should run on every prog-mode
or elisp mode buffer won't get to set a hook and
have it run. This recreates the scratch buffer
the first time it's switched to to fix this.
For some reason, simply calling the major mode
hook for the scratch buffer after doom-after-init-hook
doesn't work.