fix: treesitter mode setting itself as default
If `rust-ts-mode` is used as the base of `rust-mode`
then it will automatically add itself to `auto-mode-alist`
before `rustic-mode`, causing every file other than
the first to be opened to default to `rust-ts-mode`
instead. This can be fixed by removing the treesitter
mode from the list after it gets set when the package
loads. This will change behaviour if:

1) both this module and the tree sitter mode are enabled
2) the user loads the `rust-ts-mode` package in some way
3) the user intends for `rust-ts-mode` to be prioritized
over the doom module.

However, this is unlikely to be the case.