feat(common-lisp): make quicklisp directory configurable
Previously, `+lisp/find-file-in-quicklisp` used a hard-coded directory
aimed at a nonstandard Quicklisp install location, `~/.quicklisp`. The
default install location is `~/quicklisp`, without the dot. This commit
makes the user's Quicklisp directory configurable through the variable
`+lisp-quicklisp-directory`, which will default to `~/.quicklisp` if it
exists or to `~/quicklisp` otherwise. `+lisp/find-file-in-quicklisp` has
been updated to respect this variable.