fix(window-select): load winum earlier
This is so that a user can use `winum-auto-setup-mode-line` with value
`t`. Since that modifies the default mode line, the entire package needs
to be loaded very early for that option to be available to a user.

This is cheaper and better than mapping over every buffer and modifying
their buffer-local value for mode-line-format.