bump: :tools tree-sitter
emacs-tree-sitter/elisp-tree-sitter@3cfab8a0e945 -> emacs-tree-sitter/elisp-tree-sitter@02fe7b86d92b

tree-sitter was bumped in ee10764, but the bump was reverted by mistake
in 7719991 because the default branch in tree-sitter's MELPA recipe is
release, not master, which confused the doom/bump-package* commands.
This partially reverts 7719991 to undo the de-bump.

Amend: 7719991badf3
Ref: ee10764c220c