refactor!(org): remove org-yt
BREAKING CHANGE: This removes the org-yt package, which will break
[[yt:...]] links in org, if you're using them. The package is a source
of errors, likely because it hasn't been updated for Org's new element
parsing system.

This package was originally included for eventual use in Doom's
documentation, but it likely won't be used and isn't universally useful
enough to keep in this module by default.

Fix: #7913
Fix: #7317