feat(marginalia): Annotate `doom/help-packages'
This will provide annotations for `doom/help-packages'
command.
Specifically, it will display the source of the package:
- Straight: The package was installed using Straight (dooms package manager).
- Built-in: The package is bundled in with Emacs.
- 'Other': This is for when a package is known to Doom but not
installed (such as when you disable a package in 'package.el').

When the package is installed and was installed using Straight, then the
commit pin is displayed.

The package also displays documentation for the package (if found).