tweak: consistent order for profile init functions
Doom embeds the partials generated by functions on
`doom-profile-generators` in lexicographical order by filename, but
after recent refactoring all these partials generally do is define a
function which Doom calls from `doom-startup`. When Doom generates
`doom-startup`, it inserts these calls in `doom-profile-generators`
order.

This does not affect Doom itself, since it initializes
`doom-profile-generators` in a sensible order. But any additional
entries pushed onto `doom-profile-generators` run first, which is
surprising (and might break them if they depend on loaddefs or modules
having initialized).

Fix it by sorting the list. Do so up front just for simplicity: although
the order in which we generate the partials is unlikely to matter, if we
have to sort the list we might as well use the sorted one throughout.

Amend: b3aa41fd7424