fix(cli): some package files not being byte-compiled (2nd take)
This reverts a workaround for straight that is no longer necessary
because it has been applied upstream and we've bumped the package since.

For some reason I don't understand, this reproducibly fixes a
reoccurrence of the original problem for me: Before this commit, some
files of the citar package were not byte-compiled. After this commit,
the problem is gone.

This effectively reverts 54a084fed7e1, though it's not a reverting
commit in strict sense because the code has been moved to a different
file.

Revert: 54a084fed7e1
Ref: d80c5e51159f
Ref: #7707
Ref: radian-software/straight.el#1132
Ref: https://discourse.doomemacs.org/t/is-it-expected-that-doom-sync-does-not-byte-compile-native-compile-some-files/4946