default: add omnicompletion bindings to minibuffer
This commit adds omnicompletion bindings to the minibuffer. This is only
done for Corfu because we do not currently enable Company in the
minibuffer. This is primarily because sometimes I want to insert a file
name into the minibuffer and want to do so with C-x C-f. If you have any
suggestions on improving this, please comment below. Not all the
bindings are added, because some of them do not make sense in the
minibuffer unless you wrap them with special forms.