bump: dirvish
hlissner/dirvish@5f046190e886 -> alexluigit/dirvish@f449754dd591

As of alexluigit/dirvish@f449754dd591, all of these hacks are obsolete and can
be safely removed (see Refs below).  The hack around `:commands
dirvish-find-entry-a dirvish-dired-noselect-a` is kept as it is because I don't
understand why that is necessary.

The latest upstream also includes fixes for #8170 and #8081 (respect
`dired-kill-when-opening-new-dired-buffer` option).

Ref: alexluigit/dirvish@3145513cb517325370765949b00fbd7a991df7e2
Ref: alexluigit/dirvish@24e612c39de520b713cf8dc1027cd934daf9d28e
Ref: alexluigit/dirvish@15e2cd192591e2d01d58bca08be6fe5e4c664b82
Ref: alexluigit/dirvish@3bdb7f23b94aeebd9c6e6fd218e8135c26f5ec1b
Ref: alexluigit/dirvish@1b905092f67483bce8db4901227df15e6c548dce
Ref: alexluigit/dirvish@d775e38b530ac380e4c414db422ebfb224c95382
Ref: alexluigit/dirvish@f986123a4d3247c9c3a8186ee51d797c8fda820f
Ref: alexluigit/dirvish@0cd07f5f0958a307f95f70ecf29f10f12ff1a672
Fix: #8170
Fix: #8081