fix: check that DEBUG envvar is non-empty
Starting in 8c7711920e6dfc2996200fde786bf1b09f58730f, doom--if-compile
sets the DEBUG environment variable to an empty string if
doom-debug-mode is disabled. But in early-init.el, we were just checking
that the DEBUG envvar was present, even if DEBUG was an empty string.
Now, we check that DEBUG is non-empty. This affects commands like
doom/reload.