fix: projectile: cache advice if not in a project
c3a9388452df549e introduced a regression that breaks the following
workflow:

1. Fresh start of Emacs, no project has been switched to.

2. Press `<SPC> <SPC>`, bound to `projectile-find-file`. Since we are
not currently on a project, `projectile-find-file` does the sensible
thing and calls `projectile-switch-project`, which shows a list of
projects for us to pick one.

3. Select one such project, and subsequent presses of `<SPC> <SPC>` will
call `projectile-find-file` as expected (including the advice under
discussion), and everything is hunky-dory.

NOTE: On step 2., the advice is triggered, and it assumes that we are in
a project. Sadly, we are not yet.  So an error is produced, and shown in
the mini-buffer:

    Wrong type argument:, stringp, nil

This happens because within the advice, `(doom-project-root)` returns
`nil`.  Bummer!

This commit fixes that, by calling the original function without
arguments if `proot` is `nil`, sidestepping the issue.

Fix: c3a9388452df549e
Ref: #8317