bump: ws-butler
lewang/ws-butler@d3927f6131f2 -> emacsmirror/nongnu_elpa@9ee5a7657a22

ws-butler was moved to a branch of the emacsmirror/nongnu_elpa repo, but
Straight already uses this repo for recipe lookups and struggles to host
multiple packages that live in the same repo but different branches. To
get around this, Straight must be forced to clone it twice with an
explicit recipe.

This needs to be revisited, because this will likely happen again in the
future...

Fix: #8333
Ref: emacsmirror/nongnu_elpa@b890949b1bec