fix(vc): better line selection in git-timemachine buffers when browse at remote
`+vc/browse-at-remote` on git-timemachine buffers solely depended on the active
region to determine highlighted lines in the browser, sometimes returning an
error when no region is set, and other times using an implicit/invisible
transient mark. The number of lines that get highlighted at remote is also not
always consistent.

This commit modifies `+vc-support-git-timemachine-a` to correct all the above
and have `+vc/browse-at-remote` remain consistent between git-timemachine
buffers and non git-timemachine buffers.