fix(latex): disable smartparens pairs conflicting with AucTeX
This disables smartparens pairs for quotation marks, which are broken.
Before this commit, typing a LaTeX comment
    % "Foo"
inserted
    % ''Foo`|`''
where | indicates the position of the cursor (wtf!).

Let's better AucTeX's support for quotation marks for now and accept
that the user won't get automatic insertions of closing quotation marks.

This also removes a wrong comment mentioning the non-existing variable
`+latex-enable-plain-double-quotes'.