fix(format): prefix arg inhibiting format-on-save (part 2)
Experimenting with a new approach to this issue, in an effort to avoid
advice on such low level commands ({basic-,}save-buffer). I don't like
that we're still stacking new behavior on top of their pre-existing
prefix arg behavior (manipulating backup-on-save functionality), but
maybe it's not useful enough to worry about?

Fix: #8420
Amend: f0c8290ae2bf