refactor!(php): remove php-extras
BREAKING CHANGE: The PHP API data file no longer exists on php.net, and
there is no substitute, and php-extras is no longer maintained, so this
package is defunct and leaves php buffers in an error state. This
deprives non-LSP users from eldoc and naive code completion. The
alternative is to simply adopt LSP.